% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{act_eigfun}
\alias{act_eigfun}
\title{Apply a Function to an Operator}
\usage{
act_eigfun(object, ...)
}
\arguments{
\item{object}{a representation of the operator.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The resulting operator from the application of the function.
}
\description{
Apply a Function to an Operator
}
\seealso{
\code{\link[=act_eigfun.spectral]{act_eigfun.spectral()}}
}
