% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_evolution.R
\name{avg_matrix}
\alias{avg_matrix}
\title{Get the Average Mixing Matrix of a Quantum Walk}
\usage{
avg_matrix(object, ...)
}
\arguments{
\item{object}{a representation of the quantum walk.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The average mixing matrix.
}
\description{
Get the Average Mixing Matrix of a Quantum Walk
}
\seealso{
\code{\link[=mixing_matrix]{mixing_matrix()}}, \code{\link[=gavg_matrix]{gavg_matrix()}}, \code{\link[=avg_matrix.ctqwalk]{avg_matrix.ctqwalk()}}
}
