% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctqwalk.R
\name{ctqwalk}
\alias{ctqwalk}
\title{Create a Continuous-time Quantum Walk}
\usage{
ctqwalk(hamiltonian)
}
\arguments{
\item{hamiltonian}{a Hermitian Matrix representing the Hamiltonian of the system.}
}
\value{
An object of class "ctqwalk" containing the continuous-time quantum walk
object.
}
\description{
Function to create a continuous-time quantum walk (ctqwalk) object.
}
\examples{
# Quantum walk on the Path graph with three vertices, the hamiltonian is the adjacency matrix.
P3 <- matrix(c(0,1,0,1,0,1,0,1,0), nrow=3)

walk <- ctqwalk(P3)


}
