% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{evalMFUN}
\alias{evalMFUN}
\title{Apply a Function to a Matrix}
\usage{
evalMFUN(object, ...)
}
\arguments{
\item{object}{an object containing a representation of a matrix.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The resulting matrix from the application of the function.
}
\description{
Apply a Function to a Matrix
}
\seealso{
\link{evalMFUN.spectral}
}
