% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{extractEIGSPACE}
\alias{extractEIGSPACE}
\title{Generic S3 method extractEIGSPACE}
\usage{
extractEIGSPACE(object, ...)
}
\arguments{
\item{object}{an object containing the spectral decomposition of a matrix.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The eigenbasis of the desired eigenspace.
}
\description{
Generic S3 method extractEIGSPACE
}
\seealso{
\link{extractPROJ}, \link{extractSCHUR},
\link{extractEIGSPACE.spectral}
}
