% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{extractEIGSPACE.spectral}
\alias{extractEIGSPACE.spectral}
\title{extractEIGSPACE method for spectral objects}
\usage{
\method{extractEIGSPACE}{spectral}(object, id, ...)
}
\arguments{
\item{object}{an object of class spectral.}

\item{id}{index for the desired eigenspace according to the ordered (decreasing) spectra.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A matrix whose columns form the orthonormal eigenbasis.
}
\description{
extractEIGSPACE method for spectral objects
}
\examples{
# Spectra is {2, -1} with multiplicities one and two respectively.
decomp <- spectral(matrix(c(0,1,1,1,0,1,1,1,0), nrow=3))

# Returns the two orthonormal eigenvectors corresponding to the eigenvalue -1.
extractEIGSPACE(decomp, id=2)

# Returns the eigenvector corresponding to the eigenvalue 2.
extractEIGSPACE(decomp, id=1)

}
\seealso{
\link{spectral}, \link{extractEIGSPACE}
}
