% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{extractPROJ}
\alias{extractPROJ}
\title{Generic S3 method extractPROJ}
\usage{
extractPROJ(object, ...)
}
\arguments{
\item{object}{an object containing the spectral decomposition of a matrix.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The orthogonal projector of the desired eigenspace.
}
\description{
Generic S3 method extractPROJ
}
\seealso{
\link{extractEIGSPACE}, \link{extractSCHUR},
\link{extractPROJ.spectral}
}
