% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{extractSCHUR}
\alias{extractSCHUR}
\title{Generic S3 method extractSCHUR}
\usage{
extractSCHUR(object, ...)
}
\arguments{
\item{object}{an object containing the spectral decomposition of a matrix.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The Schur product of eigenprojectors.
}
\description{
Generic S3 method extractSCHUR
}
\seealso{
\link{extractEIGSPACE}, \link{extractSCHUR},
\link{extractSCHUR.spectral}
}
