% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_evolution.R
\name{gavg_matrix.ctqwalk}
\alias{gavg_matrix.ctqwalk}
\title{Get the Generalized Average Mixing Matrix of a Continuous-Time Quantum Walk}
\usage{
\method{gavg_matrix}{ctqwalk}(object, R, ...)
}
\arguments{
\item{object}{a representation of the quantum walk.}

\item{R}{samples from the random variable \eqn{R} (For performance, it is recommended at most 10000 samples).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The generalized average mixing matrix of the continuous-time quantum walk.

Let \eqn{M(t)} be the mixing matrix of the quantum walk and \eqn{f_R(t)} the probability
density function associated with \eqn{R}, then the generalized average mixing matrix under \eqn{R}
is defined as

\deqn{\widehat{M}_R := \mathbb{E}[M(R)] = \int_{-\infty}^{\infty} M(t)f_R(t)\textrm{d}t}

\code{gavg_matrix} returns an approximation of the above matrix by the Monte Carlo approach.
}
\description{
Get the Generalized Average Mixing Matrix of a Continuous-Time Quantum Walk
}
\examples{
walk <- ctqwalk(matrix(c(0,1,0,1,0,1,0,1,0), nrow=3))

# Return the average mixing matrix under a Standard Gaussian distribution
gavg_matrix(walk, rnorm(1000))
}
\seealso{
\code{\link[=ctqwalk]{ctqwalk()}}, \code{\link[=gavg_matrix]{gavg_matrix()}}
}
