% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{get_eigproj}
\alias{get_eigproj}
\title{Extract an Eigen-Projector from an operator}
\usage{
get_eigproj(object, ...)
}
\arguments{
\item{object}{a representation of the operator.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A representation of the requested eigen-projector.
}
\description{
Extract an Eigen-Projector from an operator
}
\examples{
s <- spectral(rbind(c(0.5, 0.3), c(0.3,0.7)))

get_eigproj(s, 1) #-> get_eigproj.spectral(...)

}
\seealso{
\code{\link[=get_eigspace]{get_eigspace()}}, \code{\link[=get_eigschur]{get_eigschur()}},
\code{\link[=get_eigproj.spectral]{get_eigproj.spectral()}}
}
