% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{J}
\alias{J}
\title{The All-Ones Matrix}
\usage{
J(n)
}
\arguments{
\item{n}{the order of the matrix.}
}
\value{
A square matrix of order \eqn{n} in which every entry is
equal to 1. The all-ones matrix is given by
\eqn{J_{n\ x\ n} = 1_{n\ x\ 1}1_{n\ x\ 1}^T}.
}
\description{
Returns the all-ones matrix of order \code{n}.
}
\examples{
# Return the all-ones matrix of order 5.
J(5)
}
\seealso{
\code{\link[=tr]{tr()}}, \code{\link[=trdot]{trdot()}}, \code{\link[=cartesian]{cartesian()}}
}
