% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_evolution.R
\name{mixing_matrix}
\alias{mixing_matrix}
\title{Get the Mixing Matrix of a Quantum Walk}
\usage{
mixing_matrix(object, ...)
}
\arguments{
\item{object}{a representation of the quantum walk.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The mixing matrix of the quantum walk.
}
\description{
Get the Mixing Matrix of a Quantum Walk
}
\seealso{
\link{unitary_matrix}, \link{avg_matrix}, \link{gavg_matrix},
\link{mixing_matrix.ctqwalk}
}
