% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctqwalk.R
\name{print.ctqwalk}
\alias{print.ctqwalk}
\title{Print the ctqwalk output}
\usage{
\method{print}{ctqwalk}(x, ...)
}
\arguments{
\item{x}{an object of the class ctqwalk.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Called mainly for its side effects. However, also
returns \code{x} invisibly.
}
\description{
Print the ctqwalk output
}
