% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwalkr-package.R
\docType{package}
\name{qwalkr-package}
\alias{qwalkr}
\alias{qwalkr-package}
\title{qwalkr: Handle Continuous-Time Quantum Walks with R}
\description{
Functions and methods to create, visualize, and investigate properties of Continous-Time Quantum Walks. In addition, tools to efficiently calculate matrices that arise in the context of Quantum Walks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vitormarquesr/qwalkr}
  \item \url{https://vitormarquesr.github.io/qwalkr/}
  \item Report bugs at \url{https://github.com/vitormarquesr/qwalkr/issues}
}

}
\author{
\strong{Maintainer}: Vitor Marques \email{vmrodriguespro@gmail.com}

}
\keyword{internal}
