% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{tr}
\alias{tr}
\title{The Trace of a Matrix}
\usage{
tr(A)
}
\arguments{
\item{A}{a square matrix.}
}
\value{
If \eqn{A} has order \eqn{n},
then \eqn{tr(A) = \sum_{i=1}^{n}a_{ii}}.
}
\description{
Computes the trace of a matrix \code{A}.
}
\examples{
A <- rbind(1:5, 2:6, 3:7)

# Calculate the trace of A
tr(A)
}
\seealso{
\code{\link[=J]{J()}}, \code{\link[=trdot]{trdot()}}, \code{\link[=cartesian]{cartesian()}}
}
