% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_evolution.R
\name{unitary_matrix}
\alias{unitary_matrix}
\title{Get the Unitary Time Evolution Operator of a Quantum Walk}
\usage{
unitary_matrix(object, ...)
}
\arguments{
\item{object}{a representation of the quantum walk.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The unitary time evolution operator.
}
\description{
Get the Unitary Time Evolution Operator of a Quantum Walk
}
\seealso{
\link{unitary_matrix.ctqwalk}
}
