% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{extractPROJ.spectral}
\alias{extractPROJ.spectral}
\title{extractPROJ method for spectral objects}
\usage{
\method{extractPROJ}{spectral}(object, id, ...)
}
\arguments{
\item{object}{an object of class spectral.}

\item{id}{index for the desired eigenspace according to the ordered (decreasing) spectra.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The orthogonal projector of the desired eigenspace. Note
that its rank is equal to the multiplicity of the eigenvalue.
}
\description{
extractPROJ method for spectral objects
}
\examples{
# Spectra is {2, -1} with multiplicities one and two respectively.
decomp <- spectral(matrix(c(0,1,1,1,0,1,1,1,0), nrow=3))

# Returns the projector associated to the eigenvalue -1.
extractPROJ(decomp, id=2)

# Returns the projector associated to the eigenvalue 2.
extractPROJ(decomp, id=1)

}
\seealso{
\link{spectral}, \link{extractPROJ}
}
