% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{extractSCHUR.spectral}
\alias{extractSCHUR.spectral}
\title{extractSCHUR method for spectral objects}
\usage{
\method{extractSCHUR}{spectral}(object, id1, id2 = NULL, ...)
}
\arguments{
\item{object}{an object of class spectral.}

\item{id1}{index for the first eigenspace according to the ordered (decreasing) spectra.}

\item{id2}{index for the second eigenspace according to the ordered (decreasing) spectra. If not provided,
it takes the same value as \code{id1}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The Schur product of the corresponding eigenprojectors.
}
\description{
extractSCHUR method for spectral objects
}
\examples{
# Spectra is {2, -1} with multiplicities one and two respectively.
decomp <- spectral(matrix(c(0,1,1,1,0,1,1,1,0), nrow=3))

# Returns the Schur product between the 2-projector and -1-projector.
extractSCHUR(decomp, id1=2, id2=1)

# Returns the Schur square of the 2-projector.
extractSCHUR(decomp, id1=1, id2=1)

# Also returns the Schur square of the 2-projector
extractSCHUR(decomp, id1=1)

}
\seealso{
\link{spectral}, \link{extractSCHUR}
}
