% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_evolution.R
\name{gavg_matrix}
\alias{gavg_matrix}
\title{The Generalized Average Mixing Matrix of a Quantum Walk}
\usage{
gavg_matrix(object, ...)
}
\arguments{
\item{object}{a representation of the quantum walk.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The generalized average mixing matrix.
}
\description{
The Generalized Average Mixing Matrix of a Quantum Walk
}
\seealso{
\code{\link[=mixing_matrix]{mixing_matrix()}}, \code{\link[=avg_matrix]{avg_matrix()}}, \code{\link[=gavg_matrix.ctqwalk]{gavg_matrix.ctqwalk()}}
}
