% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{get_eigproj}
\alias{get_eigproj}
\title{Extract an Eigen-Projector from an operator}
\usage{
get_eigproj(object, ...)
}
\arguments{
\item{object}{a representation of the operator.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A representation of the requested eigen-projector.
}
\description{
Extract an Eigen-Projector from an operator
}
\seealso{
\link{get_eigspace}, \link{get_eigschur},
\link{get_eigproj.spectral}
}
