% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{get_eigspace}
\alias{get_eigspace}
\title{Extract an Eigenspace from an Operator}
\usage{
get_eigspace(object, ...)
}
\arguments{
\item{object}{a representation of the operator.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A representation of the requested eigenspace.
}
\description{
Extract an Eigenspace from an Operator
}
\seealso{
\link{get_eigproj}, \link{extractSCHUR},
\link{get_eigspace.spectral}
}
