% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwalkr-package.R
\docType{package}
\name{qwalkr-package}
\alias{qwalkr}
\alias{qwalkr-package}
\title{qwalkr: Handle Continuous-Time Quantum Walks with R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions and tools for creating, visualizing, and investigating properties of continuous-time quantum walks, including efficient calculation of matrices such as the mixing matrix, average mixing matrix, and spectral decomposition of the Hamiltonian. E. Farhi (1997): \href{https://arxiv.org/abs/quant-ph/9706062v2}{arXiv:quant-ph/9706062v2}; C. Godsil (2011) \href{https://arxiv.org/abs/1103.2578v3}{arXiv:1103.2578v3}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vitormarquesr/qwalkr}
  \item \url{https://vitormarquesr.github.io/qwalkr/}
  \item Report bugs at \url{https://github.com/vitormarquesr/qwalkr/issues}
}

}
\author{
\strong{Maintainer}: Vitor Marques \email{vmrodriguespro@gmail.com} [copyright holder]

}
\keyword{internal}
